+++ server/src/main/kotlin/suwayomi/tachidesk/opds/dto/OpdsMangaDetails.kt
@@ -5,4 +5,5 @@ data class OpdsMangaDetails( // Kept name, it's specific enough
     val title: String,
     val thumbnailUrl: String?,
     val author: String?, // Added for chapter entry authors
+    val totalChapters: Long, // Added to determine if it's a oneshot
 )