+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Added
 - (**Sync**) Added [SyncYomi](https://github.com/syncyomi/syncyomi) support
 - (**Sync/SyncYomi**) Add trigger settings to start sync automatically on chapter read, chapter open, client start, and client resume
+- (**OPDS**) Add option to skip chapter metadata feed providing direct stream/download links
 
 ### Changed
 - (**Database/H2**) Use the latest H2 database engine
@@ -31,6 +32,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - (**API**) Fix GraphQL `Filter` `notAll` and `notAny` being inversed
 - (**API**) Fix GraphQL `Filter` causing an UnsupportedOperationException when passing an empty list as a `Any` filter value
 - (**Build**) Fix CURL failing silently in builds
+- (**Backup/Database**) Fix backup creation slowdown when mapping chapters
 
 ## [v2.2.2100] + [WebUI: v20260508.01] - 2026-05-08